some small changes; fixed command-line indexing bug, and changed lbd and bps so that they default to using devices on their own index (rather than 0).

+++ include/bpsdevice.h
@@ -110,7 +110,7 @@ class CBpsDevice : public CDevice
   private: double err;
 
   // Our index
-  private: uint16_t index;
+  private: int index;
     
   // Pointer to beacon detector to get data from
   private: CDevice *position;
